Moved plugin tests to a dedicated test file.
