+++ tests/editor/editor.js
@@ -3,11 +3,11 @@
  * For licensing, see LICENSE.md.
  */
 
-/* globals describe, it, expect, beforeEach, sinon, document, setTimeout */
+/* globals describe, it, expect, beforeEach, sinon, document */
 
 'use strict';
 
-var modules = bender.amd.require( 'editor', 'editorconfig', 'plugin', 'promise' );
+var modules = bender.amd.require( 'editor', 'editorconfig', 'promise' );
 
 var editor;
 var element;
@@ -20,60 +20,6 @@ beforeEach( function() {
 
 	editor = new Editor( element );
 } );
-
-// Define fake plugins to be used in tests.
-
-CKEDITOR.define( 'plugin!A', [ 'plugin' ], function( Plugin ) {
-	return Plugin.extend( {
-		init: sinon.spy().named( 'A' )
-	} );
-} );
-
-CKEDITOR.define( 'plugin!B', [ 'plugin' ], function( Plugin ) {
-	return Plugin.extend( {
-		init: sinon.spy().named( 'B' )
-	} );
-} );
-
-CKEDITOR.define( 'plugin!C', [ 'plugin', 'plugin!B' ], function( Plugin ) {
-	return Plugin.extend( {
-		init: sinon.spy().named( 'C' )
-	} );
-} );
-
-CKEDITOR.define( 'plugin!D', [ 'plugin', 'plugin!C' ], function( Plugin ) {
-	return Plugin.extend( {
-		init: sinon.spy().named( 'D' )
-	} );
-} );
-
-CKEDITOR.define( 'plugin!E', [ 'plugin' ], function( Plugin ) {
-	return Plugin.extend( {} );
-} );
-
-// Synchronous plugin that depends on an asynchronous one.
-CKEDITOR.define( 'plugin!F', [ 'plugin', 'plugin!async' ], function( Plugin ) {
-	return Plugin.extend( {
-		init: sinon.spy().named( 'F' )
-	} );
-} );
-
-var asyncSpy = sinon.spy().named( 'async-call-spy' );
-
-CKEDITOR.define( 'plugin!async', [ 'plugin', 'promise' ], function( Plugin, Promise ) {
-	return Plugin.extend( {
-		init: sinon.spy( function() {
-			return new Promise( function( resolve ) {
-				setTimeout( function() {
-					asyncSpy();
-					resolve();
-				}, 0 );
-			} );
-		} )
-	} );
-} );
-
-///////////////////
 
 describe( 'constructor', function() {
 	it( 'should create a new editor instance', function() {
@@ -105,74 +51,6 @@ describe( 'init', function() {
 
 		expect( editor.init() ).to.equal( promise );
 	} );
-
-	it( 'should fill `plugins`', function() {
-		var Editor = modules.editor;
-		var Plugin = modules.plugin;
-
-		editor = new Editor( element, {
-			plugins: 'A,B'
-		} );
-
-		expect( editor.plugins.length ).to.equal( 0 );
-
-		return editor.init().then( function() {
-			expect( editor.plugins.length ).to.equal( 2 );
-
-			expect( editor.plugins.get( 'A' ) ).to.be.an.instanceof( Plugin );
-			expect( editor.plugins.get( 'B' ) ).to.be.an.instanceof( Plugin );
-		} );
-	} );
-
-	it( 'should initialize plugins in the right order', function() {
-		var Editor = modules.editor;
-
-		editor = new Editor( element, {
-			plugins: 'A,D'
-		} );
-
-		return editor.init().then( function() {
-			sinon.assert.callOrder(
-				editor.plugins.get( 'A' ).init,
-				editor.plugins.get( 'B' ).init,
-				editor.plugins.get( 'C' ).init,
-				editor.plugins.get( 'D' ).init
-			);
-		} );
-	} );
-
-	it( 'should initialize plugins in the right order, waiting for asynchronous ones', function() {
-		var Editor = modules.editor;
-
-		editor = new Editor( element, {
-			plugins: 'A,F'
-		} );
-
-		return editor.init().then( function() {
-			sinon.assert.callOrder(
-				editor.plugins.get( 'A' ).init,
-				editor.plugins.get( 'async' ).init,
-				asyncSpy,	// This one is called with delay by the async init
-				editor.plugins.get( 'F' ).init
-			);
-		} );
-	} );
-
-	it( 'should not fail if loading a plugin that doesn\'t define init()', function() {
-		var Editor = modules.editor;
-
-		editor = new Editor( element, {
-			plugins: 'E'
-		} );
-
-		return editor.init();
-	} );
-} );
-
-describe( 'plugins', function() {
-	it( 'should be empty on new editor', function() {
-		expect( editor.plugins.length ).to.equal( 0 );
-	} );
 } );
 
 describe( 'destroy', function() {
@@ -181,14 +59,14 @@ describe( 'destroy', function() {
 
 		editor.on( 'destroy', spy );
 
-		editor.destroy();
-
-		sinon.assert.called( spy );
+		return editor.destroy().then( function() {
+			sinon.assert.called( spy );
+		} );
 	} );
 
-	it( 'should delete the "element" property', function() {
-		editor.destroy();
-
-		expect( editor ).to.not.have.property( 'element' );
+	it( 'should undefine the "element" property', function() {
+		return editor.destroy().then( function() {
+			expect( editor ).to.not.have.property( 'element' );
+		} );
 	} );
 } );
